requirements.txt
streamlit>=1.37.0
pandas>=2.2.2
numpy>=1.26.4
plotly>=5.23.0
reportlab>=4.2.2
requests>=2.32.3
python-dotenv>=1.0.1
